feat(terminal): auto-focus newly created terminal tab
Extracts the double-rAF xterm focus helper into a shared module and calls
it from all four new-terminal entry points (Cmd+T, "+" menu, IPC-triggered,
and per-group newTerminalTab/newTerminalWithShell). Previously only the
"+" menu path focused the new tab, so keyboard and IPC-created terminals
would drop the first keystroke.
